Fix windows test

+++ tests/IResearch/IResearchLinkTest.cpp
@@ -2520,7 +2520,7 @@ TEST_F(IResearchLinkMetricsTest, WriteAndMetrics1) {
   }
   {
     auto cid = static_cast<unsigned long long>(_logicalCollection->id().id());
-    auto iid = _link->id().id();
+    auto iid = static_cast<unsigned long long>(_link->id().id());
     char expectedData[1000];  // clang-format off
     std::sprintf(expectedData,
       "# HELP arangodb_search_num_docs Number of documents\n"
@@ -2595,7 +2595,7 @@ TEST_F(IResearchLinkMetricsTest, WriteAndMetrics2) {
   }
   {
     auto cid = static_cast<unsigned long long>(_logicalCollection->id().id());
-    auto iid = _link->id().id();
+    auto iid = static_cast<unsigned long long>(_link->id().id());
     char expectedData[1000];  // clang-format off
     std::sprintf(expectedData,
       "# HELP arangodb_search_num_docs Number of documents\n"
@@ -2634,7 +2634,7 @@ TEST_F(IResearchLinkMetricsTest, WriteAndMetrics2) {
   }
   {
     auto cid = static_cast<unsigned long long>(_logicalCollection->id().id());
-    auto iid = _link->id().id();
+    auto iid = static_cast<unsigned long long>(_link->id().id());
     char expectedData[1000];  // clang-format off
     std::sprintf(expectedData,
       "# HELP arangodb_search_num_docs Number of documents\n"
@@ -2673,7 +2673,7 @@ TEST_F(IResearchLinkMetricsTest, LinkAndMetics) {
     insert(1, 2, 0);
 
     auto cid = static_cast<unsigned long long>(_logicalCollection->id().id());
-    auto iid = _link->id().id();
+    auto iid = static_cast<unsigned long long>(_link->id().id());
     char expectedData[1000];  // clang-format off
     std::sprintf(expectedData,
       "# HELP arangodb_search_num_docs Number of documents\n"
@@ -2701,7 +2701,7 @@ TEST_F(IResearchLinkMetricsTest, LinkAndMetics) {
     insert(2, 3, 2);
 
     auto cid = static_cast<unsigned long long>(_logicalCollection->id().id());
-    auto iid = _link->id().id();
+    auto iid = static_cast<unsigned long long>(_link->id().id());
     char expectedData[1000];  // clang-format off
     std::sprintf(expectedData,
       "# HELP arangodb_search_num_docs Number of documents\n"
